@@ -2,6 +2,7 @@ import os
 import re
 import sys
 import warnings
+from colorsys import hls_to_rgb, rgb_to_hls
 from itertools import cycle, combinations
 from functools import partial
 from typing import Callable, List, Union
@@ -9,12 +10,14 @@ from typing import Callable, List, Union
 import numpy as np
 import pandas as pd
 
+from bokeh.colors import RGB
 from bokeh.colors.named import (
     lime as BULL_COLOR,
     tomato as BEAR_COLOR
 )
 from bokeh.plotting import figure as _figure
 from bokeh.models import (
+    CrosshairTool,
     CustomJS,
     ColumnDataSource,
     NumeralTickFormatter,
@@ -80,19 +83,20 @@ def colorgen():
 
 
 def lightness(color, lightness=.94):
-    color = color.to_hsl()
-    color.l = lightness  # noqa
-    return color.to_rgb()
+    rgb = np.array([color.r, color.g, color.b]) / 255
+    h, _, s = rgb_to_hls(*rgb)
+    rgb = np.array(hls_to_rgb(h, lightness, s)) * 255
+    return RGB(*rgb)
 
 
-_MAX_CANDLES = 10000
+_MAX_CANDLES = 10_000
 
 
 def _maybe_resample_data(resample_rule, df, indicators, equity_data, trades):
     if isinstance(resample_rule, str):
         freq = resample_rule
     else:
-        if len(df) < _MAX_CANDLES:
+        if resample_rule is False or len(df) <= _MAX_CANDLES:
             return df, indicators, equity_data, trades
 
         from_index = dict(day=-2, hour=-6, minute=1, second=0, millisecond=0,
@@ -100,8 +104,8 @@ def _maybe_resample_data(resample_rule, df, indicators, equity_data, trades):
         FREQS = ('1T', '5T', '10T', '15T', '30T', '1H', '2H', '4H', '8H', '1D', '1W', '1M')
         freq = next((f for f in FREQS[from_index:]
                      if len(df.resample(f)) <= _MAX_CANDLES), FREQS[-1])
-        warnings.warn("Data contains too many candlesticks to plot; downsampling to {!r}. "
-                      "See `Backtest.plot(resample=...)`".format(freq))
+        warnings.warn(f"Data contains too many candlesticks to plot; downsampling to {freq!r}. "
+                      "See `Backtest.plot(resample=...)`")
 
     from .lib import OHLCV_AGG, TRADES_AGG, _EQUITY_AGG
     df = df.resample(freq, label='right').agg(OHLCV_AGG).dropna()
@@ -109,8 +113,8 @@ def _maybe_resample_data(resample_rule, df, indicators, equity_data, trades):
     indicators = [_Indicator(i.df.resample(freq, label='right').mean()
                              .dropna().reindex(df.index).values.T,
                              **dict(i._opts, name=i.name,
-                                    # HACK: override `data` for its index
-                                    data=pd.Series(np.nan, index=df.index)))
+                                    # Replace saved index with the resampled one
+                                    index=df.index))
                   for i in indicators]
     assert not indicators or indicators[0].df.index.equals(df.index)
 
@@ -146,7 +150,7 @@ def plot(*, results: pd.Series,
          df: pd.DataFrame,
          indicators: List[_Indicator],
          filename='', plot_width=None,
-         plot_equity=True, plot_pl=True,
+         plot_equity=True, plot_return=False, plot_pl=True,
          plot_volume=True, plot_drawdown=False,
          smooth_equity=False, relative_equity=True,
          superimpose=True, resample=True,
@@ -170,7 +174,10 @@ def plot(*, results: pd.Series,
     trades = results['_trades']
 
     plot_volume = plot_volume and not df.Volume.isnull().all()
-    is_datetime_index = df.index.is_all_dates
+    plot_equity = plot_equity and not trades.empty
+    plot_return = plot_return and not trades.empty
+    plot_pl = plot_pl and not trades.empty
+    is_datetime_index = isinstance(df.index, pd.DatetimeIndex)
 
     from .lib import OHLCV_AGG
     # ohlc df may contain many columns. We're only interested in, and pass on to Bokeh, these
@@ -192,7 +199,7 @@ def plot(*, results: pd.Series,
         x_axis_type='linear',
         plot_width=plot_width,
         plot_height=400,
-        tools="xpan,xwheel_zoom,box_zoom,undo,redo,reset,crosshair,save",
+        tools="xpan,xwheel_zoom,box_zoom,undo,redo,reset,save",
         active_drag='xpan',
         active_scroll='xwheel_zoom')
 
@@ -200,6 +207,7 @@ def plot(*, results: pd.Series,
 
     fig_ohlc = new_bokeh_figure(
         x_range=Range1d(index[0], index[-1],
+                        min_interval=10,
                         bounds=(index[0] - pad,
                                 index[-1] + pad)) if index.size > 1 else None)
     figs_above_ohlc, figs_below_ohlc = [], []
@@ -234,7 +242,7 @@ this.labels = this.labels || formatter.doFormat(ticks
 return this.labels[index] || "";
         ''')
 
-    NBSP = '&nbsp;' * 4
+    NBSP = '\N{NBSP}' * 4
     ohlc_extreme_values = df[['High', 'Low']].copy(deep=False)
     ohlc_tooltips = [
         ('x, y', NBSP.join(('$index',
@@ -255,7 +263,7 @@ return this.labels[index] || "";
         fig.yaxis.minor_tick_line_color = None
         return fig
 
-    def set_tooltips(fig, tooltips=(), vline=True, renderers=(), show_arrow=True):
+    def set_tooltips(fig, tooltips=(), vline=True, renderers=()):
         tooltips = list(tooltips)
         renderers = list(renderers)
 
@@ -267,10 +275,10 @@ return this.labels[index] || "";
             tooltips = [("#", "@index")] + tooltips
         fig.add_tools(HoverTool(
             point_policy='follow_mouse',
-            renderers=renderers, formatters=formatters, show_arrow=show_arrow,
+            renderers=renderers, formatters=formatters,
             tooltips=tooltips, mode='vline' if vline else 'mouse'))
 
-    def _plot_equity_section():
+    def _plot_equity_section(is_return=False):
         """Equity section"""
         # Max DD Dur. line
         equity = equity_data['Equity'].copy()
@@ -294,7 +302,7 @@ return this.labels[index] || "";
                 # Include max dd end points. Otherwise the MaxDD line looks amiss.
                 dd_start, int(dd_end), min(int(dd_end + 1), equity.size - 1),
             ])
-            select = pd.Index(trades['ExitBar']) | interest_points
+            select = pd.Index(trades['ExitBar']).union(interest_points)
             select = select.unique().dropna()
             equity = equity.iloc[select].reindex(equity.index)
             equity.interpolate(inplace=True)
@@ -303,10 +311,14 @@ return this.labels[index] || "";
 
         if relative_equity:
             equity /= equity.iloc[0]
+        if is_return:
+            equity -= equity.iloc[0]
 
-        source.add(equity, 'equity')
+        yaxis_label = 'Return' if is_return else 'Equity'
+        source_key = 'eq_return' if is_return else 'equity'
+        source.add(equity, source_key)
         fig = new_indicator_figure(
-            y_axis_label="Equity",
+            y_axis_label=yaxis_label,
             **({} if plot_drawdown else dict(plot_height=110)))
 
         # High-watermark drawdown dents
@@ -318,16 +330,16 @@ return this.labels[index] || "";
                   fill_color='#ffffea', line_color='#ffcb66')
 
         # Equity line
-        r = fig.line('index', 'equity', source=source, line_width=1.5, line_alpha=1)
+        r = fig.line('index', source_key, source=source, line_width=1.5, line_alpha=1)
         if relative_equity:
-            tooltip_format = '@equity{+0,0.[000]%}'
+            tooltip_format = f'@{source_key}{{+0,0.[000]%}}'
             tick_format = '0,0.[00]%'
             legend_format = '{:,.0f}%'
         else:
-            tooltip_format = '@equity{$ 0,0}'
+            tooltip_format = f'@{source_key}{{$ 0,0}}'
             tick_format = '$ 0.0 a'
             legend_format = '${:,.0f}'
-        set_tooltips(fig, [('Equity', tooltip_format)], renderers=[r])
+        set_tooltips(fig, [(yaxis_label, tooltip_format)], renderers=[r])
         fig.yaxis.formatter = NumeralTickFormatter(format=tick_format)
 
         # Peaks
@@ -350,7 +362,7 @@ return this.labels[index] || "";
         dd_timedelta_label = df['datetime'].iloc[int(round(dd_end))] - df['datetime'].iloc[dd_start]
         fig.line([dd_start, dd_end], equity.iloc[dd_start],
                  line_color='red', line_width=2,
-                 legend_label='Max Dd Dur. ({})'.format(dd_timedelta_label)
+                 legend_label=f'Max Dd Dur. ({dd_timedelta_label})'
                  .replace(' 00:00:00', '')
                  .replace('(0 days ', '('))
 
@@ -420,8 +432,8 @@ return this.labels[index] || "";
                               millisecond='S').get(time_resolution))
         if not resample_rule:
             warnings.warn(
-                "'Can't superimpose OHLC data with rule '{}' (index datetime resolution: '{}'). "
-                "Skipping.".format(resample_rule, time_resolution),
+                f"'Can't superimpose OHLC data with rule '{resample_rule}'"
+                f"(index datetime resolution: '{time_resolution}'). Skipping.",
                 stacklevel=4)
             return
 
@@ -465,7 +477,7 @@ return this.labels[index] || "";
         trade_source.add(trades[['EntryPrice', 'ExitPrice']].values.tolist(), 'position_lines_ys')
         fig_ohlc.multi_line(xs='position_lines_xs', ys='position_lines_ys',
                             source=trade_source, line_color=trades_cmap,
-                            legend_label='Trades ({})'.format(len(trades)),
+                            legend_label=f'Trades ({len(trades)})',
                             line_width=8, line_alpha=1, line_dash='dotted')
 
     def _plot_indicators():
@@ -474,7 +486,7 @@ return this.labels[index] || "";
         def _too_many_dims(value):
             assert value.ndim >= 2
             if value.ndim > 2:
-                warnings.warn("Can't plot indicators with >2D ('{}')".format(value.name),
+                warnings.warn(f"Can't plot indicators with >2D ('{value.name}')",
                               stacklevel=5)
                 return True
             return False
@@ -513,11 +525,11 @@ return this.labels[index] || "";
             legend_label = LegendStr(value.name)
             for j, arr in enumerate(value, 1):
                 color = next(colors)
-                source_name = '{}_{}_{}'.format(legend_label, i, j)
+                source_name = f'{legend_label}_{i}_{j}'
                 if arr.dtype == bool:
                     arr = arr.astype(int)
                 source.add(arr, source_name)
-                tooltips.append('@{{{}}}{{0,0.0[0000]}}'.format(source_name))
+                tooltips.append(f'@{{{source_name}}}{{0,0.0[0000]}}')
                 if is_overlay:
                     ohlc_extreme_values[source_name] = arr
                     if is_scatter:
@@ -565,6 +577,9 @@ return this.labels[index] || "";
     if plot_equity:
         _plot_equity_section()
 
+    if plot_return:
+        _plot_equity_section(is_return=True)
+
     if plot_drawdown:
         figs_above_ohlc.append(_plot_drawdown_section())
 
@@ -599,21 +614,26 @@ return this.labels[index] || "";
                                                   code=_AUTOSCALE_JS_CALLBACK))
 
     plots = figs_above_ohlc + [fig_ohlc] + figs_below_ohlc
+    linked_crosshair = CrosshairTool(dimensions='both')
+
     for f in plots:
         if f.legend:
-            f.legend.location = 'top_left' if show_legend else None
+            f.legend.visible = show_legend
+            f.legend.location = 'top_left'
             f.legend.border_line_width = 1
             f.legend.border_line_color = '#333333'
             f.legend.padding = 5
             f.legend.spacing = 0
             f.legend.margin = 0
             f.legend.label_text_font_size = '8pt'
+            f.legend.click_policy = "hide"
         f.min_border_left = 0
         f.min_border_top = 3
         f.min_border_bottom = 6
         f.min_border_right = 10
         f.outline_line_color = '#666666'
 
+        f.add_tools(linked_crosshair)
         wheelzoom_tool = next(wz for wz in f.tools if isinstance(wz, WheelZoomTool))
         wheelzoom_tool.maintain_focus = False
 
