@@ -4,11 +4,6 @@ Objects from this module can also be imported from the top-level
 module directly, e.g.
 
     from backtesting import Backtest, Strategy
-
-.. warning:: v0.2.0 breaking changes
-   Version 0.2.0 introduced some **breaking API changes**. For quick ways to
-   migrate existing 0.1.x code, see the implementing
-   [pull request](https://github.com/kernc/backtesting.py/pull/47/).
 """
 import multiprocessing as mp
 import os
@@ -17,8 +12,8 @@ import warnings
 from abc import abstractmethod, ABCMeta
 from concurrent.futures import ProcessPoolExecutor, as_completed
 from copy import copy
-from functools import partial
-from itertools import repeat, product, chain
+from functools import lru_cache, partial
+from itertools import repeat, product, chain, compress
 from math import copysign
 from numbers import Number
 from typing import Callable, Dict, List, Optional, Sequence, Tuple, Type, Union
@@ -54,27 +49,27 @@ class Strategy(metaclass=ABCMeta):
     """
     def __init__(self, broker, data, params):
         self._indicators = []
-        self._broker = broker  # type: _Broker
-        self._data = data   # type: _Data
+        self._broker: _Broker = broker
+        self._data: _Data = data
         self._params = self._check_params(params)
 
     def __repr__(self):
         return '<Strategy ' + str(self) + '>'
 
     def __str__(self):
-        params = ','.join('{}={}'.format(*p) for p in zip(self._params.keys(),
-                                                          map(_as_str, self._params.values())))
+        params = ','.join(f'{i[0]}={i[1]}' for i in zip(self._params.keys(),
+                                                        map(_as_str, self._params.values())))
         if params:
             params = '(' + params + ')'
-        return '{}{}'.format(self.__class__.__name__, params)
+        return f'{self.__class__.__name__}{params}'
 
     def _check_params(self, params):
         for k, v in params.items():
             if not hasattr(self, k):
                 raise AttributeError(
-                    "Strategy '{}' is missing parameter '{}'. Strategy class "
-                    "should define parameters as class variables before they "
-                    "can be optimized or run with.".format(self.__class__.__name__, k))
+                    f"Strategy '{self.__class__.__name__}' is missing parameter '{k}'."
+                    "Strategy class should define parameters as class variables before they "
+                    "can be optimized or run with.")
             setattr(self, k, v)
         return params
 
@@ -121,7 +116,7 @@ class Strategy(metaclass=ABCMeta):
         if name is None:
             params = ','.join(filter(None, map(_as_str, chain(args, kwargs.values()))))
             func_name = _as_str(func)
-            name = ('{}({})' if params else '{}').format(func_name, params)
+            name = (f'{func_name}({params})' if params else f'{func_name}')
         else:
             name = name.format(*map(_as_str, args),
                                **dict(zip(kwargs.keys(), map(_as_str, kwargs.values()))))
@@ -129,12 +124,13 @@ class Strategy(metaclass=ABCMeta):
         try:
             value = func(*args, **kwargs)
         except Exception as e:
-            raise RuntimeError('Indicator "{}" errored with exception: {}'.format(name, e))
+            raise RuntimeError(f'Indicator "{name}" errored with exception: {e}')
 
         if isinstance(value, pd.DataFrame):
             value = value.values.T
 
-        value = try_(lambda: np.asarray(value, order='C'), None)
+        if value is not None:
+            value = try_(lambda: np.asarray(value, order='C'), None)
         is_arraylike = value is not None
 
         # Optionally flip the array if the user returned e.g. `df.values`
@@ -144,8 +140,8 @@ class Strategy(metaclass=ABCMeta):
         if not is_arraylike or not 1 <= value.ndim <= 2 or value.shape[-1] != len(self._data.Close):
             raise ValueError(
                 'Indicators must return (optionally a tuple of) numpy.arrays of same '
-                'length as `data`(data shape: {}; indicator "{}" shape: {}, value: {})'
-                .format(self._data.Close.shape, name, getattr(value, 'shape', ''), value))
+                f'length as `data` (data shape: {self._data.Close.shape}; indicator "{name}"'
+                f'shape: {getattr(value, "shape" , "")}, returned value: {value})')
 
         if plot and overlay is None and np.issubdtype(value.dtype, np.number):
             x = value / self._data.Close
@@ -157,7 +153,7 @@ class Strategy(metaclass=ABCMeta):
         value = _Indicator(value, name=name, plot=plot, overlay=overlay,
                            color=color, scatter=scatter,
                            # _Indicator.s Series accessor uses this:
-                           data=self.data)
+                           index=self.data.index)
         self._indicators.append(value)
         return value
 
@@ -192,8 +188,12 @@ class Strategy(metaclass=ABCMeta):
             super().next()
         """
 
+    class __FULL_EUITY(float):
+        def __repr__(self): return '.9999'
+    _FULL_EUITY = __FULL_EUITY(1 - sys.float_info.epsilon)
+
     def buy(self, *,
-            size: float = 1 - sys.float_info.epsilon,
+            size: float = _FULL_EUITY,
             limit: float = None,
             stop: float = None,
             sl: float = None,
@@ -208,7 +208,7 @@ class Strategy(metaclass=ABCMeta):
         return self._broker.new_order(size, limit, stop, sl, tp)
 
     def sell(self, *,
-             size: float = 1 - sys.float_info.epsilon,
+             size: float = _FULL_EUITY,
              limit: float = None,
              stop: float = None,
              sl: float = None,
@@ -292,10 +292,10 @@ class _Orders(tuple):
         removed_attrs = ('entry', 'set_entry', 'is_long', 'is_short',
                          'sl', 'tp', 'set_sl', 'set_tp')
         if item in removed_attrs:
-            raise AttributeError('Strategy.orders.{} were removed in Backtesting 0.2.0. '
-                                 'Use `Order` API instead. See docs.'
-                                 .format('/.'.join(removed_attrs)))
-        raise AttributeError("'tuple' object has no attribute {!r}".format(item))
+            raise AttributeError(f'Strategy.orders.{"/.".join(removed_attrs)} were removed in'
+                                 'Backtesting 0.2.0. '
+                                 'Use `Order` API instead. See docs.')
+        raise AttributeError(f"'tuple' object has no attribute {item!r}")
 
 
 class Position:
@@ -350,7 +350,7 @@ class Position:
             trade.close(portion)
 
     def __repr__(self):
-        return '<Position: {} ({} trades)>'.format(self.size, len(self.__broker.trades))
+        return f'<Position: {self.size} ({len(self.__broker.trades)} trades)>'
 
 
 class _OutOfMoneyError(Exception):
@@ -390,11 +390,11 @@ class Order:
 
     def _replace(self, **kwargs):
         for k, v in kwargs.items():
-            setattr(self, '_{}__{}'.format(self.__class__.__qualname__, k), v)
+            setattr(self, f'_{self.__class__.__qualname__}__{k}', v)
         return self
 
     def __repr__(self):
-        return '<Order {}>'.format(', '.join('{}={}'.format(param, round(value, 5))
+        return '<Order {}>'.format(', '.join(f'{param}={round(value, 5)}'
                                              for param, value in (
                                                  ('size', self.__size),
                                                  ('limit', self.__limit_price),
@@ -510,22 +510,21 @@ class Trade:
         self.__broker = broker
         self.__size = size
         self.__entry_price = entry_price
-        self.__exit_price = None  # type: Optional[float]
-        self.__entry_bar = entry_bar  # type: int
-        self.__exit_bar = None  # type: Optional[int]
-        self.__sl_order = None  # type: Optional[Order]
-        self.__tp_order = None  # type: Optional[Order]
+        self.__exit_price: Optional[float] = None
+        self.__entry_bar: int = entry_bar
+        self.__exit_bar: Optional[int] = None
+        self.__sl_order: Optional[Order] = None
+        self.__tp_order: Optional[Order] = None
         self.__exit_type = "Saida por max" # type: Optional[String]
         self.__entry_rule = 0 # type: Optional[int]
-   
+
     def __repr__(self):
-        return '<Trade size={} time={}-{} price={}-{} pl={:.0f} entry={} exit={}>'.format(
-            self.__size, self.__entry_bar, self.__exit_bar or '',
-            self.__entry_price, self.__exit_price or '', self.pl, self.__entry_rule,  self.__exit_type)
+        return f'<Trade size={self.__size} time={self.__entry_bar}-{self.__exit_bar or ""} ' \
+               f'price={self.__entry_price}-{self.__exit_price or ""} pl={self.pl:.0f} entry={self.__entry_rule} exit={self.__exit_type}>'
 
     def _replace(self, **kwargs):
         for k, v in kwargs.items():
-            setattr(self, '_{}__{}'.format(self.__class__.__qualname__, k), v)
+            setattr(self, f'_{self.__class__.__qualname__}__{k}', v)
         return self
 
     def _copy(self, **kwargs):
@@ -668,8 +667,8 @@ class Trade:
     def __set_contingent(self, type, price):
         assert type in ('sl', 'tp')
         assert price is None or 0 < price < np.inf
-        attr = '_{}__{}_order'.format(self.__class__.__qualname__, type)
-        order = getattr(self, attr)  # type: Order
+        attr = f'_{self.__class__.__qualname__}__{type}_order'
+        order: Order = getattr(self, attr)
         if order:
             order.cancel()
         if price:
@@ -681,10 +680,10 @@ class Trade:
 class _Broker:
     def __init__(self, *, data, cash, commission, margin,
                  trade_on_close, hedging, exclusive_orders, index, tradeOnHit):
-        assert 0 < cash, "cash shosuld be >0, is {}".format(cash)
-        assert 0 <= commission < .1, "commission should be between 0-10%, is {}".format(commission)
-        assert 0 < margin <= 1, "margin should be between 0 and 1, is {}".format(margin)
-        self._data = data  # type: _Data
+        assert 0 < cash, f"cash should be >0, is {cash}"
+        assert 0 <= commission < .1, f"commission should be between 0-10%, is {commission}"
+        assert 0 < margin <= 1, f"margin should be between 0 and 1, is {margin}"
+        self._data: _Data = data
         self._cash = cash
         self._commission = commission
         self._leverage = 1 / margin
@@ -694,14 +693,13 @@ class _Broker:
         self._trade_on_hit = tradeOnHit
 
         self._equity = np.tile(np.nan, len(index))
-        self.orders = []  # type: List[Order]
-        self.trades = []  # type: List[Trade]
+        self.orders: List[Order] = []
+        self.trades: List[Trade] = []
         self.position = Position(self)
-        self.closed_trades = []  # type: List[Trade]
+        self.closed_trades: List[Trade] = []
 
     def __repr__(self):
-        return '<Broker: {:.0f}{:+.1f} ({} trades)>'.format(
-            self._cash, self.position.pl, len(self.trades))
+        return f'<Broker: {self._cash:.0f}{self.position.pl:+.1f} ({len(self.trades)} trades)>'
 
     def new_order(self,
                   size: float,
@@ -721,15 +719,18 @@ class _Broker:
         tp = tp and float(tp)
 
         is_long = size > 0
+        adjusted_price = self._adjusted_price(size)
 
         if is_long:
-            if not (sl or -np.inf) <= (limit or stop or self.last_price) <= (tp or np.inf):
-                raise ValueError("Long orders require: SL ({}) < LIMIT ({}) < TP ({})".format(
-                    sl, limit or stop or self.last_price, tp))
+            if not (sl or -np.inf) < (limit or stop or adjusted_price) < (tp or np.inf):
+                raise ValueError(
+                    "Long orders require: "
+                    f"SL ({sl}) < LIMIT ({limit or stop or adjusted_price}) < TP ({tp})")
         else:
-            if not (tp or -np.inf) <= (limit or stop or self.last_price) <= (sl or np.inf):
-                raise ValueError("Short orders require: TP ({}) < LIMIT ({}) < SL ({})".format(
-                    tp, limit or stop or self.last_price, sl))
+            if not (tp or -np.inf) < (limit or stop or adjusted_price) < (sl or np.inf):
+                raise ValueError(
+                    "Short orders require: "
+                    f"TP ({tp}) < LIMIT ({limit or stop or adjusted_price}) < SL ({sl})")
 
         order = Order(self, size, limit, stop, sl, tp, trade)
         # Put the new order in the order queue,
@@ -752,10 +753,15 @@ class _Broker:
 
     @property
     def last_price(self) -> float:
-        """Return price at the last (current) close.
-        Used e.g. in `Orders._is_price_ok()` to see if the set price is reasonable.
-        """
+        """ Price at the last (current) close. """
         return self._data.Close[-1]
+
+    def _adjusted_price(self, size=None, price=None) -> float:
+        """
+        Long/short `price`, adjusted for commisions.
+        In long positions, the adjusted price is a fraction higher, and vice versa.
+        """
+        return (price or self.last_price) * (1 + copysign(self._commission, size))
 
     @property
     def equity(self) -> float:
@@ -831,9 +837,9 @@ class _Broker:
                         continue
 
                 # stop_price, if set, was hit within this bar
-                price = (min(open, order.limit, stop_price or np.inf)
+                price = (min(stop_price or open, order.limit)
                          if order.is_long else
-                         max(open, order.limit, stop_price or -np.inf))
+                         max(stop_price or open, order.limit))
             else:
                 # Market-if-touched / market order
                 price = prev_close if self._trade_on_close else open
@@ -849,9 +855,12 @@ class _Broker:
             if order.parent_trade:
                 trade = order.parent_trade
                 _prev_size = trade.size
+                # If order.size is "greater" than trade.size, this order is a trade.close()
+                # order and part of the trade was already closed beforehand
+                size = copysign(min(abs(_prev_size), abs(order.size)), order.size)
                 # If this trade isn't already closed (e.g. on multiple `trade.close(.5)` calls)
                 if trade in self.trades:
-                    self._reduce_trade(trade, price, order.size, time_index)
+                    self._reduce_trade(trade, price, size, time_index)
                     assert order.size != -_prev_size or trade not in self.trades
                 if order in (trade._sl_order,
                              trade._tp_order):
@@ -859,7 +868,7 @@ class _Broker:
                     assert order not in self.orders  # Removed when trade was closed
                 else:
                     # It's a trade.close() order, now done
-                    assert abs(_prev_size) >= abs(order.size) >= 1
+                    assert abs(_prev_size) >= abs(size) >= 1
                     self.orders.remove(order)
                 continue
 
@@ -867,7 +876,7 @@ class _Broker:
 
             # Adjust price to include commission (or bid-ask spread).
             # In long positions, the adjusted price is a fraction higher, and vice versa.
-            adjusted_price = price * (1 + copysign(self._commission, order.size))
+            adjusted_price = self._adjusted_price(order.size, price)
 
             # If order size was specified proportionally,
             # precompute true size in units, accounting for margin and spread/commissions
@@ -901,6 +910,8 @@ class _Broker:
                         # so it will only be closed partially
                         self._reduce_trade(trade, price, need_size, time_index)
                         need_size = 0
+
+                    if not need_size:
                         break
 
             # If we don't have enough liquidity to cover for the order, cancel it
@@ -921,9 +932,10 @@ class _Broker:
                     elif (low <= (order.sl or -np.inf) <= high or
                           low <= (order.tp or -np.inf) <= high):
                         warnings.warn(
-                            "A SL/TP order would execute in the same bar as its contingent upon "
-                            "stop/limit order. Since we can't assert the precise intra-candle "
-                            "price movement, the affected SL/TP order will be executed on "
+                            f"({data.index[-1]}) A contingent SL/TP order would execute in the "
+                            "same bar its parent stop/limit order was turned into a trade. "
+                            "Since we can't assert the precise intra-candle "
+                            "price movement, the affected SL/TP order will instead be executed on "
                             "the next (matching) price/bar, making the result (of this trade) "
                             "somewhat dubious. "
                             "See https://github.com/kernc/backtesting.py/issues/119",
@@ -940,6 +952,7 @@ class _Broker:
         assert abs(trade.size) >= abs(size)
 
         size_left = trade.size + size
+        assert size_left * trade.size >= 0
         if not size_left:
             close_trade = trade
         else:
@@ -993,7 +1006,7 @@ class Backtest:
                  data: pd.DataFrame,
                  strategy: Type[Strategy],
                  *,
-                 cash: float = 10000,
+                 cash: float = 10_000,
                  commission: float = .0,
                  margin: float = 1.,
                  trade_on_close=False,
@@ -1060,7 +1073,7 @@ class Backtest:
         data = data.copy(deep=False)
 
         # Convert index to datetime index
-        if (not data.index.is_all_dates and
+        if (not isinstance(data.index, pd.DatetimeIndex) and
             not isinstance(data.index, pd.RangeIndex) and
             # Numeric index with most large numbers
             (data.index.is_numeric() and
@@ -1075,23 +1088,28 @@ class Backtest:
 
         if len(data) == 0:
             raise ValueError('OHLC `data` is empty')
-        if len(data.columns & {'Open', 'High', 'Low', 'Close', 'Volume'}) != 5:
+        if len(data.columns.intersection({'Open', 'High', 'Low', 'Close', 'Volume'})) != 5:
             raise ValueError("`data` must be a pandas.DataFrame with columns "
                              "'Open', 'High', 'Low', 'Close', and (optionally) 'Volume'")
         if data[['Open', 'High', 'Low', 'Close']].isnull().values.any():
             raise ValueError('Some OHLC values are missing (NaN). '
                              'Please strip those lines with `df.dropna()` or '
                              'fill them in with `df.interpolate()` or whatever.')
+        if np.any(data['Close'] > cash):
+            warnings.warn('Some prices are larger than initial cash value. Note that fractional '
+                          'trading is not supported. If you want to trade Bitcoin, '
+                          'increase initial cash, or trade μBTC or satoshis instead (GH-134).',
+                          stacklevel=2)
         if not data.index.is_monotonic_increasing:
             warnings.warn('Data index is not sorted in ascending order. Sorting.',
                           stacklevel=2)
             data = data.sort_index()
-        if not data.index.is_all_dates:
+        if not isinstance(data.index, pd.DatetimeIndex):
             warnings.warn('Data index is not datetime. Assuming simple periods, '
                           'but `pd.DateTimeIndex` is advised.',
                           stacklevel=2)
 
-        self._data = data   # type: pd.DataFrame
+        self._data: pd.DataFrame = data
         self._broker = partial(
             _Broker, cash=cash, commission=commission, margin=margin,
             trade_on_close=trade_on_close, hedging=hedging,
@@ -1118,6 +1136,11 @@ class Backtest:
             Equity Peak [$]                       75787.4
             Return [%]                            419.599
             Buy & Hold Return [%]                 703.458
+            Return (Ann.) [%]                      21.328
+            Volatility (Ann.) [%]                 36.5383
+            Sharpe Ratio                         0.583718
+            Sortino Ratio                         1.09239
+            Calmar Ratio                         0.444518
             Max. Drawdown [%]                    -47.9801
             Avg. Drawdown [%]                    -5.92585
             Max. Drawdown Duration      584 days 00:00:00
@@ -1132,17 +1155,14 @@ class Backtest:
             Profit Factor                         2.08802
             Expectancy [%]                        8.79171
             SQN                                  0.916893
-            Sharpe Ratio                         0.179141
-            Sortino Ratio                         0.55887
-            Calmar Ratio                         0.049056
             _strategy                            SmaCross
             _equity_curve                           Eq...
             _trades                       Size  EntryB...
             dtype: object
         """
         data = _Data(self._data.copy(deep=False))
-        broker = self._broker(data=data)  # type: _Broker
-        strategy = self._strategy(broker, data, kwargs)  # type: Strategy
+        broker: _Broker = self._broker(data=data)
+        strategy: Strategy = self._strategy(broker, data, kwargs)
 
         strategy.init()
         data._update()  # Strategy.init might have changed/added to data.df
@@ -1176,29 +1196,61 @@ class Backtest:
 
                 # Next tick, a moment before bar close
                 strategy.next()
+            else:
+                # Close any remaining open trades so they produce some stats
+                for trade in broker.trades:
+                    trade.close()
 
-        # Set data back to full length
-        # for future `indicator._opts['data'].index` calls to work
-        data._set_length(len(self._data))
+                # Re-run broker one last time to handle orders placed in the last strategy
+                # iteration. Use the same OHLC values as in the last broker iteration.
+                if start < len(self._data):
+                    try_(broker.next, exception=_OutOfMoneyError)
 
-        self._results = self._compute_stats(broker, strategy)
+            # Set data back to full length
+            # for future `indicator._opts['data'].index` calls to work
+            data._set_length(len(self._data))
+
+            self._results = self._compute_stats(broker, strategy)
         return self._results
 
-    def optimize(self,
+    def optimize(self, *,
                  maximize: Union[str, Callable[[pd.Series], float]] = 'SQN',
+                 method: str = 'grid',
+                 max_tries: Union[int, float] = None,
                  constraint: Callable[[dict], bool] = None,
                  return_heatmap: bool = False,
-                 **kwargs) -> Union[pd.Series, Tuple[pd.Series, pd.Series]]:
+                 return_optimization: bool = False,
+                 random_state: int = None,
+                 **kwargs) -> Union[pd.Series,
+                                    Tuple[pd.Series, pd.Series],
+                                    Tuple[pd.Series, pd.Series, dict]]:
         """
-        Optimize strategy parameters to an optimal combination using
-        parallel exhaustive search. Returns result `pd.Series` of
-        the best run.
+        Optimize strategy parameters to an optimal combination.
+        Returns result `pd.Series` of the best run.
 
         `maximize` is a string key from the
         `backtesting.backtesting.Backtest.run`-returned results series,
         or a function that accepts this series object and returns a number;
         the higher the better. By default, the method maximizes
         Van Tharp's [System Quality Number](https://google.com/search?q=System+Quality+Number).
+
+        `method` is the optimization method. Currently two methods are supported:
+
+        * `"grid"` which does an exhaustive (or randomized) search over the
+          cartesian product of parameter combinations, and
+        * `"skopt"` which finds close-to-optimal strategy parameters using
+          [model-based optimization], making at most `max_tries` evaluations.
+
+        [model-based optimization]: \
+            https://scikit-optimize.github.io/stable/auto_examples/bayesian-optimization.html
+
+        `max_tries` is the maximal number of strategy runs to perform.
+        If `method="grid"`, this results in randomized grid search.
+        If `max_tries` is a floating value between (0, 1], this sets the
+        number of runs to approximately that fraction of full grid space.
+        Alternatively, if integer, it denotes the absolute maximum number
+        of evaluations. If unspecified (default), grid search is exhaustive,
+        whereas for `method="skopt"`, `max_tries` is set to 200.
 
         `constraint` is a function that accepts a dict-like object of
         parameters (with values) and returns `True` when the combination
@@ -1211,6 +1263,20 @@ class Backtest:
         inspected or projected onto 2D to plot a heatmap
         (see `backtesting.lib.plot_heatmaps()`).
 
+        If `return_optimization` is True and `method = 'skopt'`,
+        in addition to result series (and maybe heatmap), return raw
+        [`scipy.optimize.OptimizeResult`][OptimizeResult] for further
+        inspection, e.g. with [scikit-optimize]\
+        [plotting tools].
+
+        [OptimizeResult]: \
+            https://docs.scipy.org/doc/scipy/reference/generated/scipy.optimize.OptimizeResult.html
+        [scikit-optimize]: https://scikit-optimize.github.io
+        [plotting tools]: https://scikit-optimize.github.io/stable/modules/plots.html
+
+        If you want reproducible optimization results, set `random_state`
+        to a fixed integer or a `numpy.random.RandomState` object.
+
         Additional keyword arguments represent strategy arguments with
         list-like collections of possible values. For example, the following
         code finds and returns the "best" of the 7 admissible (of the
@@ -1218,10 +1284,6 @@ class Backtest:
 
             backtest.optimize(sma1=[5, 10, 15], sma2=[10, 20, 40],
                               constraint=lambda p: p.sma1 < p.sma2)
-
-        .. TODO::
-            Add parameter `max_tries: Union[int, float] = None` which switches
-            from exhaustive grid search to random search. See notes in the source.
 
         .. TODO::
             Improve multiprocessing/parallel execution on Windos with start method 'spawn'.
@@ -1245,6 +1307,7 @@ class Backtest:
                             'Series) or a function that accepts result Series '
                             'and returns a number; the higher the better')
 
+        have_constraint = bool(constraint)
         if constraint is None:
 
             def constraint(_):
@@ -1255,83 +1318,199 @@ class Backtest:
                             "of strategy parameters and returns a bool whether "
                             "the combination of parameters is admissible or not")
 
+        if return_optimization and method != 'skopt':
+            raise ValueError("return_optimization=True only valid if method='skopt'")
+
         def _tuple(x):
             return x if isinstance(x, Sequence) and not isinstance(x, str) else (x,)
+
+        for k, v in kwargs.items():
+            if len(_tuple(v)) == 0:
+                raise ValueError(f"Optimization variable '{k}' is passed no "
+                                 f"optimization values: {k}={v}")
 
         class AttrDict(dict):
             def __getattr__(self, item):
                 return self[item]
 
-        param_combos = tuple(map(dict,  # back to dict so it pickles
-                                 filter(constraint,  # constraints applied on our fancy dict
-                                        map(AttrDict,
-                                            product(*(zip(repeat(k), _tuple(v))
-                                                      for k, v in kwargs.items()))))))
-        if not param_combos:
-            raise ValueError('No admissible parameter combinations to test')
+        def _grid_size():
+            size = np.prod([len(_tuple(v)) for v in kwargs.values()])
+            if size < 10_000 and have_constraint:
+                size = sum(1 for p in product(*(zip(repeat(k), _tuple(v))
+                                                for k, v in kwargs.items()))
+                           if constraint(AttrDict(p)))
+            return size
 
-        if len(param_combos) > 300:
-            warnings.warn('Searching for best of {} configurations.'.format(len(param_combos)),
-                          stacklevel=2)
+        def _optimize_grid() -> Union[pd.Series, Tuple[pd.Series, pd.Series]]:
+            rand = np.random.RandomState(random_state).random
+            grid_frac = (1 if max_tries is None else
+                         max_tries if 0 < max_tries <= 1 else
+                         max_tries / _grid_size())
+            param_combos = [dict(params)  # back to dict so it pickles
+                            for params in (AttrDict(params)
+                                           for params in product(*(zip(repeat(k), _tuple(v))
+                                                                   for k, v in kwargs.items())))
+                            if constraint(params)  # type: ignore
+                            and rand() <= grid_frac]
+            if not param_combos:
+                raise ValueError('No admissible parameter combinations to test')
 
-        heatmap = pd.Series(np.nan,
-                            name=maximize_key,
-                            index=pd.MultiIndex.from_tuples([p.values() for p in param_combos],
-                                                            names=next(iter(param_combos)).keys()))
+            if len(param_combos) > 300:
+                warnings.warn(f'Searching for best of {len(param_combos)} configurations.',
+                              stacklevel=2)
 
-        # TODO: add parameter `max_tries:Union[int, float]=None` which switches
-        # exhaustive grid search to random search. This might need to avoid
-        # returning NaNs in stats on runs with no trades to differentiate those
-        # from non-tested parameter combos in heatmap.
+            heatmap = pd.Series(np.nan,
+                                name=maximize_key,
+                                index=pd.MultiIndex.from_tuples(
+                                    [p.values() for p in param_combos],
+                                    names=next(iter(param_combos)).keys()))
 
-        def _batch(seq):
-            n = np.clip(len(seq) // (os.cpu_count() or 1), 5, 300)
-            for i in range(0, len(seq), n):
-                yield seq[i:i + n]
+            def _batch(seq):
+                n = np.clip(int(len(seq) // (os.cpu_count() or 1)), 1, 300)
+                for i in range(0, len(seq), n):
+                    yield seq[i:i + n]
 
-        # Save necessary objects into "global" state; pass into concurrent executor
-        # (and thus pickle) nothing but two numbers; receive nothing but numbers.
-        # With start method "fork", children processes will inherit parent address space
-        # in a copy-on-write manner, achieving better performance/RAM benefit.
-        backtest_uuid = np.random.random()
-        param_batches = list(_batch(param_combos))
-        Backtest._mp_backtests[backtest_uuid] = (self, param_batches, maximize)  # type: ignore
-        try:
-            # If multiprocessing start method is 'fork' (i.e. on POSIX), use
-            # a pool of processes to compute results in parallel.
-            # Otherwise (i.e. on Windos), sequential computation will be "faster".
-            if mp.get_start_method(allow_none=False) == 'fork':
-                with ProcessPoolExecutor() as executor:
-                    futures = [executor.submit(Backtest._mp_task, backtest_uuid, i)
-                               for i in range(len(param_batches))]
-                    for future in _tqdm(as_completed(futures), total=len(futures)):
-                        batch_index, values = future.result()
+            # Save necessary objects into "global" state; pass into concurrent executor
+            # (and thus pickle) nothing but two numbers; receive nothing but numbers.
+            # With start method "fork", children processes will inherit parent address space
+            # in a copy-on-write manner, achieving better performance/RAM benefit.
+            backtest_uuid = np.random.random()
+            param_batches = list(_batch(param_combos))
+            Backtest._mp_backtests[backtest_uuid] = (self, param_batches, maximize)  # type: ignore
+            try:
+                # If multiprocessing start method is 'fork' (i.e. on POSIX), use
+                # a pool of processes to compute results in parallel.
+                # Otherwise (i.e. on Windos), sequential computation will be "faster".
+                if mp.get_start_method(allow_none=False) == 'fork':
+                    with ProcessPoolExecutor() as executor:
+                        futures = [executor.submit(Backtest._mp_task, backtest_uuid, i)
+                                   for i in range(len(param_batches))]
+                        for future in _tqdm(as_completed(futures), total=len(futures),
+                                            desc='Backtest.optimize'):
+                            batch_index, values = future.result()
+                            for value, params in zip(values, param_batches[batch_index]):
+                                heatmap[tuple(params.values())] = value
+                else:
+                    if os.name == 'posix':
+                        warnings.warn("For multiprocessing support in `Backtest.optimize()` "
+                                      "set multiprocessing start method to 'fork'.")
+                    for batch_index in _tqdm(range(len(param_batches))):
+                        _, values = Backtest._mp_task(backtest_uuid, batch_index)
                         for value, params in zip(values, param_batches[batch_index]):
                             heatmap[tuple(params.values())] = value
+            finally:
+                del Backtest._mp_backtests[backtest_uuid]
+
+            best_params = heatmap.idxmax()
+
+            if pd.isnull(best_params):
+                # No trade was made in any of the runs. Just make a random
+                # run so we get some, if empty, results
+                stats = self.run(**param_combos[0])
             else:
-                if os.name == 'posix':
-                    warnings.warn("For multiprocessing support in `Backtest.optimize()` "
-                                  "set multiprocessing start method to 'fork'.")
-                for batch_index in _tqdm(range(len(param_batches))):
-                    _, values = Backtest._mp_task(backtest_uuid, batch_index)
-                    for value, params in zip(values, param_batches[batch_index]):
-                        heatmap[tuple(params.values())] = value
-        finally:
-            del Backtest._mp_backtests[backtest_uuid]
+                stats = self.run(**dict(zip(heatmap.index.names, best_params)))
 
-        best_params = heatmap.idxmax()
+            if return_heatmap:
+                return stats, heatmap
+            return stats
 
-        if pd.isnull(best_params):
-            # No trade was made in any of the runs. Just make a random
-            # run so we get some, if empty, results
-            self.run(**param_combos[0])  # type: ignore
+        def _optimize_skopt() -> Union[pd.Series,
+                                       Tuple[pd.Series, pd.Series],
+                                       Tuple[pd.Series, pd.Series, dict]]:
+            try:
+                from skopt import forest_minimize
+                from skopt.space import Integer, Real, Categorical
+                from skopt.utils import use_named_args
+                from skopt.callbacks import DeltaXStopper
+                from skopt.learning import ExtraTreesRegressor
+            except ImportError:
+                raise ImportError("Need package 'scikit-optimize' for method='skopt'. "
+                                  "pip install scikit-optimize")
+
+            nonlocal max_tries
+            max_tries = (200 if max_tries is None else
+                         max(1, int(max_tries * _grid_size())) if 0 < max_tries <= 1 else
+                         max_tries)
+
+            dimensions = []
+            for key, values in kwargs.items():
+                values = np.asarray(values)
+                if values.dtype.kind in 'mM':  # timedelta, datetime64
+                    # these dtypes are unsupported in skopt, so convert to raw int
+                    # TODO: save dtype and convert back later
+                    values = values.astype(int)
+
+                if values.dtype.kind in 'iumM':
+                    dimensions.append(Integer(low=values.min(), high=values.max(), name=key))
+                elif values.dtype.kind == 'f':
+                    dimensions.append(Real(low=values.min(), high=values.max(), name=key))
+                else:
+                    dimensions.append(Categorical(values.tolist(), name=key, transform='onehot'))
+
+            # Avoid recomputing re-evaluations:
+            # "The objective has been evaluated at this point before."
+            # https://github.com/scikit-optimize/scikit-optimize/issues/302
+            memoized_run = lru_cache()(lambda tup: self.run(**dict(tup)))
+
+            # np.inf/np.nan breaks sklearn, np.finfo(float).max breaks skopt.plots.plot_objective
+            INVALID = 1e300
+            progress = iter(_tqdm(repeat(None), total=max_tries, desc='Backtest.optimize'))
+
+            @use_named_args(dimensions=dimensions)
+            def objective_function(**params):
+                next(progress)
+                # Check constraints
+                # TODO: Adjust after https://github.com/scikit-optimize/scikit-optimize/pull/971
+                if not constraint(AttrDict(params)):
+                    return INVALID
+                res = memoized_run(tuple(params.items()))
+                value = -maximize(res)
+                if np.isnan(value):
+                    return INVALID
+                return value
+
+            with warnings.catch_warnings():
+                warnings.filterwarnings(
+                    'ignore', 'The objective has been evaluated at this point before.')
+
+                res = forest_minimize(
+                    func=objective_function,
+                    dimensions=dimensions,
+                    n_calls=max_tries,
+                    base_estimator=ExtraTreesRegressor(n_estimators=20, min_samples_leaf=2),
+                    acq_func='LCB',
+                    kappa=3,
+                    n_initial_points=min(max_tries, 20 + 3 * len(kwargs)),
+                    initial_point_generator='lhs',  # 'sobel' requires n_initial_points ~ 2**N
+                    callback=DeltaXStopper(9e-7),
+                    random_state=random_state)
+
+            stats = self.run(**dict(zip(kwargs.keys(), res.x)))
+            output = [stats]
+
+            if return_heatmap:
+                heatmap = pd.Series(dict(zip(map(tuple, res.x_iters), -res.func_vals)),
+                                    name=maximize_key)
+                heatmap.index.names = kwargs.keys()
+                heatmap = heatmap[heatmap != -INVALID]
+                heatmap.sort_index(inplace=True)
+                output.append(heatmap)
+
+            if return_optimization:
+                valid = res.func_vals != INVALID
+                res.x_iters = list(compress(res.x_iters, valid))
+                res.func_vals = res.func_vals[valid]
+                output.append(res)
+
+            return stats if len(output) == 1 else tuple(output)
+
+        if method == 'grid':
+            output = _optimize_grid()
+        elif method == 'skopt':
+            output = _optimize_skopt()
         else:
-            # Re-run best strategy so that the next .plot() call will render it
-            self.run(**dict(zip(heatmap.index.names, best_params)))
-
-        if return_heatmap:
-            return self._results, heatmap
-        return self._results
+            raise ValueError(f"Method should be 'grid' or 'skopt', not {method!r}")
+        return output
 
     @staticmethod
     def _mp_task(backtest_uuid, batch_index):
@@ -1340,7 +1519,7 @@ class Backtest:
                              for stats in (bt.run(**params)
                                            for params in param_batches[batch_index])]
 
-    _mp_backtests = {}  # type: Dict[float, Tuple[Backtest, List, Callable]]
+    _mp_backtests: Dict[float, Tuple['Backtest', List, Callable]] = {}
 
     @staticmethod
     def _compute_drawdown_duration_peaks(dd: pd.Series):
@@ -1363,7 +1542,7 @@ class Backtest:
         equity = pd.Series(broker._equity).bfill().fillna(broker._cash).values
         dd = 1 - equity / np.maximum.accumulate(equity)
         ddBox = 1 - (equity-(equity[0] - self._boxSize)) / np.maximum.accumulate((equity-(equity[0] - self._boxSize)))
-        dd_dur, dd_peaks = self._compute_drawdown_duration_peaks(pd.Series(dd, index=data.index))
+        dd_dur, dd_peaks = self._compute_drawdown_duration_peaks(pd.Series(dd, index=index))
 
         equity_df = pd.DataFrame({
             'Equity': equity,
@@ -1433,26 +1612,55 @@ class Backtest:
         s.loc['Max. Drawdown [%]'] = max_dd = -np.nan_to_num(dd.max()) * 100
         s.loc['Max. Drawdown [$]'] = -(np.nan_to_num(dd.max()) * np.maximum.accumulate(equity)).max()
         s.loc['Max. Box Drawdown [%]'] = -np.nan_to_num(ddBox.max()) * 100
+        s.loc['Buy & Hold Return [%]'] = (c[-1] - c[0]) / c[0] * 100  # long-only return
+
+        def geometric_mean(returns):
+            returns = returns.fillna(0) + 1
+            return (0 if np.any(returns <= 0) else
+                    np.exp(np.log(returns).sum() / (len(returns) or np.nan)) - 1)
+
+        day_returns = gmean_day_return = np.array(np.nan)
+        annual_trading_days = np.nan
+        if isinstance(index, pd.DatetimeIndex):
+            day_returns = equity_df['Equity'].resample('D').last().dropna().pct_change()
+            gmean_day_return = geometric_mean(day_returns)
+            annual_trading_days = float(
+                365 if index.dayofweek.to_series().between(5, 6).mean() > 2/7 * .6 else
+                252)
+
+        # Annualized return and risk metrics are computed based on the (mostly correct)
+        # assumption that the returns are compounded. See: https://dx.doi.org/10.2139/ssrn.3054517
+        # Our annualized return matches `empyrical.annual_return(day_returns)` whereas
+        # our risk doesn't; they use the simpler approach below.
+        annualized_return = (1 + gmean_day_return)**annual_trading_days - 1
+        s.loc['Return (Ann.) [%]'] = annualized_return * 100
+        s.loc['Volatility (Ann.) [%]'] = np.sqrt((day_returns.var(ddof=int(bool(day_returns.shape))) + (1 + gmean_day_return)**2)**annual_trading_days - (1 + gmean_day_return)**(2*annual_trading_days)) * 100  # noqa: E501
+        # s.loc['Return (Ann.) [%]'] = gmean_day_return * annual_trading_days * 100
+        # s.loc['Risk (Ann.) [%]'] = day_returns.std(ddof=1) * np.sqrt(annual_trading_days) * 100
+
+        # Our Sharpe mismatches `empyrical.sharpe_ratio()` because they use arithmetic mean return
+        # and simple standard deviation
+        s.loc['Sharpe Ratio'] = np.clip(s.loc['Return (Ann.) [%]'] / (s.loc['Volatility (Ann.) [%]'] or np.nan), 0, np.inf)  # noqa: E501
+        # Our Sortino mismatches `empyrical.sortino_ratio()` because they use arithmetic mean return
+        s.loc['Sortino Ratio'] = np.clip(annualized_return / (np.sqrt(np.mean(day_returns.clip(-np.inf, 0)**2)) * np.sqrt(annual_trading_days)), 0, np.inf)  # noqa: E501
+        max_dd = -np.nan_to_num(dd.max())
+        s.loc['Calmar Ratio'] = np.clip(annualized_return / (-max_dd or np.nan), 0, np.inf)
+        s.loc['Max. Drawdown [%]'] = max_dd * 100
         s.loc['Avg. Drawdown [%]'] = -dd_peaks.mean() * 100
         s.loc['Max. Drawdown Duration'] = _round_timedelta(dd_dur.max())
         s.loc['Avg. Drawdown Duration'] = _round_timedelta(dd_dur.mean())
         s.loc['# Trades'] = n_trades = len(trades)
         s.loc['# Comissions'] = (-1.0) * (len(trades)*self._fixedCommision)
-
-        s.loc['Win Rate [%]'] = win_rate = np.nan if not n_trades else (pl > 0).sum() / n_trades * 100  # noqa: E501
+        s.loc['Win Rate [%]'] = np.nan if not n_trades else (pl > 0).sum() / n_trades * 100  # noqa: E501
         s.loc['Best Trade [%]'] = returns.max() * 100
         s.loc['Worst Trade [%]'] = returns.min() * 100
-        mean_return = np.exp(np.log(1 + returns).sum() / (len(returns) or np.nan)) - 1
+        mean_return = geometric_mean(returns)
         s.loc['Avg. Trade [%]'] = mean_return * 100
         s.loc['Max. Trade Duration'] = _round_timedelta(durations.max())
         s.loc['Avg. Trade Duration'] = _round_timedelta(durations.mean())
         s.loc['Profit Factor'] = returns[returns > 0].sum() / (abs(returns[returns < 0].sum()) or np.nan)  # noqa: E501
-        s.loc['Expectancy [%]'] = ((returns[returns > 0].mean() * win_rate -
-                                    returns[returns < 0].mean() * (100 - win_rate)))
+        s.loc['Expectancy [%]'] = returns.mean() * 100
         s.loc['SQN'] = np.sqrt(n_trades) * pl.mean() / (pl.std() or np.nan)
-        s.loc['Sharpe Ratio'] = mean_return / (returns.std() or np.nan)
-        s.loc['Sortino Ratio'] = mean_return / (returns[returns < 0].std() or np.nan)
-        s.loc['Calmar Ratio'] = mean_return / ((-max_dd / 100) or np.nan)
         s.loc['Calmar Ratio alt'] = s.loc['Yearly Return Box[%]']  / ((-s.loc['Max. Box Drawdown [%]'] ) or np.nan)
 
         s.loc['_strategy'] = strategy
@@ -1469,7 +1677,7 @@ class Backtest:
                 return super().__repr__()
 
     def plot(self, *, results: pd.Series = None, filename=None, plot_width=None,
-             plot_equity=True, plot_pl=True,
+             plot_equity=True, plot_return=False, plot_pl=True,
              plot_volume=True, plot_drawdown=False,
              smooth_equity=False, relative_equity=True,
              superimpose: Union[bool, str] = True,
@@ -1493,7 +1701,12 @@ class Backtest:
         currently non-adjustable.
 
         If `plot_equity` is `True`, the resulting plot will contain
-        an equity (cash plus assets) graph section.
+        an equity (initial cash plus assets) graph section. This is the same
+        as `plot_return` plus initial 100%.
+
+        If `plot_return` is `True`, the resulting plot will contain
+        a cumulative return graph section. This is the same
+        as `plot_equity` minus initial 100%.
 
         If `plot_pl` is `True`, the resulting plot will contain
         a profit/loss (P/L) indicator section.
@@ -1558,6 +1771,7 @@ class Backtest:
             filename=filename,
             plot_width=plot_width,
             plot_equity=plot_equity,
+            plot_return=plot_return,
             plot_pl=plot_pl,
             plot_volume=plot_volume,
             plot_drawdown=plot_drawdown,
