@@ -1,11 +1,4 @@
 """
-# Backtesting.py Documentation
-
-.. warning:: v0.2.0 breaking changes
-   Version 0.2.0 introduced some **breaking API changes**. For quick ways to
-   migrate existing 0.1.x code, see the implementing
-   [pull request](https://github.com/kernc/backtesting.py/pull/47/).
-
 ## Manuals
 
 * [**Quick Start User Guide**](../examples/Quick Start User Guide.html)
@@ -14,24 +7,35 @@
 
 * [Library of Utilities and Composable Base Strategies](../examples/Strategies Library.html)
 * [Multiple Time Frames](../examples/Multiple Time Frames.html)
-* [Parameter Heatmap](../examples/Parameter Heatmap.html)
-* [**Trading with Machine Learning**](../examples/Trading with Machine Learning.html)
+* [**Parameter Heatmap & Optimization**](../examples/Parameter Heatmap &amp; Optimization.html)
+* [Trading with Machine Learning](../examples/Trading with Machine Learning.html)
 
-These tutorials are also available to test as live Jupyter notebooks:
+These tutorials are also available as live Jupyter notebooks:
 [![Binder](https://mybinder.org/badge_logo.svg)][binder]
+[![Google Colab](https://colab.research.google.com/assets/colab-badge.svg)][colab]
+<br>In Colab, you might have to `!pip install backtesting`.
 
 [binder]: \
     https://mybinder.org/v2/gh/kernc/backtesting.py/master?\
 urlpath=lab%2Ftree%2Fdoc%2Fexamples%2FQuick%20Start%20User%20Guide.ipynb
+[colab]: https://colab.research.google.com/github/kernc/backtesting.py/
 
 ## Example Strategies
 
 * (contributions welcome)
 
+
+.. tip::
+    For an overview of recent changes, see
+    [What's New](https://github.com/kernc/backtesting.py/blob/master/CHANGELOG.md).
+
+
 ## FAQ
 
-Potentially outdated answers to popular questions can be found on the
-[issue tracker](https://github.com/kernc/backtesting.py/issues?q=label%3Aquestion).
+Some answers to frequent and popular questions can be found on the
+[issue tracker](https://github.com/kernc/backtesting.py/issues?q=label%3Aquestion+-label%3Ainvalid)
+or on the [discussion forum](https://github.com/kernc/backtesting.py/discussions) on GitHub.
+Please use the search!
 
 ## License
 
@@ -48,7 +52,7 @@ itself find their way back to the community.
 try:
     from ._version import version as __version__  # noqa: F401
 except ImportError:
-    pass  # Package not installed
+    __version__ = '?.?.?'  # Package not installed
 
 from .backtesting import Backtest, Strategy  # noqa: F401
 from . import lib  # noqa: F401
