@@ -6,7 +6,7 @@
 #       extension: .py
 #       format_name: light
 #       format_version: '1.5'
-#       jupytext_version: 1.5.1
+#       jupytext_version: 1.6.0
 #   kernelspec:
 #     display_name: Python 3
 #     language: python
@@ -86,7 +86,12 @@ class Sma4Cross(Strategy):
 
 # -
 
-# It's not a robust strategy, but we can optimize it. Let's optimize our strategy on Google stock data.
+# It's not a robust strategy, but we can optimize it.
+#
+# [Grid search](https://en.wikipedia.org/wiki/Hyperparameter_optimization#Grid_search)
+# is an exhaustive search through a set of specified sets of values of hyperparameters. One evaluates the performance for each set of parameters and finally selects the combination that performs best.
+#
+# Let's optimize our strategy on Google stock data using _randomized_ grid search over the parameter space, evaluating at most (approximately) 200 randomly chosen combinations:
 
 # +
 # %%time 
@@ -104,13 +109,15 @@ stats, heatmap = backtest.optimize(
     n_exit=range(10, 25, 5),
     constraint=lambda p: p.n_exit < p.n_enter < p.n1 < p.n2,
     maximize='Equity Final [$]',
+    max_tries=200,
+    random_state=0,
     return_heatmap=True)
 # -
 
 # Notice `return_heatmap=True` parameter passed to
 # [`Backtest.optimize()`](https://kernc.github.io/backtesting.py/doc/backtesting/backtesting.html#backtesting.backtesting.Backtest.optimize).
 # It makes the function return a heatmap series along with the usual stats of the best run.
-# `heatmap` is a pandas Series indexed with a MultiIndex, a cartesian product of all permissible parameter values.
+# `heatmap` is a pandas Series indexed with a MultiIndex, a cartesian product of all permissible (tried) parameter values.
 # The series values are from the `maximize=` argument we provided.
 
 heatmap
@@ -151,6 +158,59 @@ from backtesting.lib import plot_heatmaps
 
 
 plot_heatmaps(heatmap, agg='mean')
+# -
+
+# ## Model-based optimization
+#
+# Above, we used _randomized grid search_ optimization method. Any kind of grid search, however, might be computationally expensive for large data sets. In the follwing example, we will use
+# [_scikit-optimize_](https://scikit-optimize.github.io)
+# package to guide our optimization better informed using forests of decision trees.
+# The hyperparameter model is sequentially improved by evaluating the expensive function (the backtest) at the next best point, thereby hopefully converging to a set of optimal parameters with as few evaluations as possible.
+#
+# So, with `method="skopt"`:
+
+# +
+# %%capture
+
+# ! pip install scikit-optimize  # This is a run-time dependency
+
+# +
+# %%time
+
+stats_skopt, heatmap, optimize_result = backtest.optimize(
+    n1=[10, 100],      # Note: For method="skopt", we
+    n2=[20, 200],      # only need interval end-points
+    n_enter=[10, 40],
+    n_exit=[10, 30],
+    constraint=lambda p: p.n_exit < p.n_enter < p.n1 < p.n2,
+    maximize='Equity Final [$]',
+    method='skopt',
+    max_tries=200,
+    random_state=0,
+    return_heatmap=True,
+    return_optimization=True)
+# -
+
+heatmap.sort_values().iloc[-3:]
+
+# Notice how the optimization runs somewhat slower even though `max_tries=` is the same. But that's due to the sequential nature of the algorithm and should actually perform rather comparably even in cases of _much larger parameter spaces_ where grid search would effectively blow up, but likely (hopefully) reaching a better local optimum than a randomized search would.
+# A note of warning, again, to take steps to avoid
+# [overfitting](https://en.wikipedia.org/wiki/Overfitting)
+# insofar as possible.
+#
+# Understanding the impact of each parameter on the computed objective function is easy in two dimensions, but as the number of dimensions grows, partial dependency plots are increasingly useful.
+# [Plotting tools from _scikit-optimize_](https://scikit-optimize.github.io/stable/modules/plots.html)
+# take care of many of the more mundane things needed to make good and informative plots of the parameter space:
+
+# +
+from skopt.plots import plot_objective
+
+_ = plot_objective(optimize_result, n_points=10)
+
+# +
+from skopt.plots import plot_evaluations
+
+_ = plot_evaluations(optimize_result, bins=10)
 # -
 
 # Learn more by exploring further
